complete module creation task
