+++ js/mock.js
@@ -1,0 +1,51 @@
+const NAMES = [
+  'Иван',
+  'Хуан',
+  'Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+  'Карл',
+];
+
+const SURNAMES = [
+  'Крутой',
+  'Ворона',
+  'Мирабелла',
+  'Шольц',
+  'Петров',
+  'Уж',
+  'Жигулевский',
+  'Ирвинг',
+  'Кукуруза',
+  'Вино',
+];
+
+const DESCRIPTIONS = [
+  'Чудесный день!',
+  'Всем добра!',
+  'Пост здесь ради лайков!',
+  'Не проходим мимо, подписываемся!',
+  'Жизнь прекрасна, выпьем Шомпанского!',
+  'Наконец-то дошли лапы выложить фотки с отпуска...',
+];
+
+const USER_COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const USERS_ARTICLES = 25;
+
+export {NAMES};
+export {SURNAMES};
+export {DESCRIPTIONS};
+export {USER_COMMENTS};
+export {USERS_ARTICLES};
